Add a signal handler for SIGUSR2.
Summary: For being able to notify (/waking) a sleeping daemon, a signal handler for the SIGUSR2 is added to the `DaemonOverseer`. This signal will be send to the child process which can implement a signal handler.

Test Plan: Put a PhutilDaemon (with a signal handler on SIGUSR2) to sleep for 10 minutes. Meanwhile, send a SIGURS2 to the daemon and see it being waked up.

Reviewers: epriestley, #blessed_reviewers

Reviewed By: epriestley

CC: Korvin, epriestley, aran, richardvanvelzen

Differential Revision: https://secure.phabricator.com/D7676