Use new Google+ API for Google auth instead of "userinfo"
Summary:
Google is moving away from the APIs we're using, as described in this timetable:

  https://developers.google.com/+/api/auth-migration#timetable

Although the timetable says the API is good until Sept 1, 2014, we're seeing the API fail on `secure.phabricator.com` and have received user reports of failures as well.

I can't figure out any configuration changes that users can make to keep the current APIs working, so just upgrade to the new ones.

Test Plan: Registered and logged in with Google.

Reviewers: btrahan

Reviewed By: btrahan

Subscribers: charles, epriestley

Differential Revision: https://secure.phabricator.com/D8612