Treat non-string arguments as normal arguments in PhutilArgumentParser
Summary:
This treats non-string arguments as "normal" arguments in `PhutilArgumentParser` (basically it prevents it from trying to detect `-` or `--` prefixes).

In Omni, I use the PhutilArgumentParser to parse arguments, but Omni supports passing structured data (arrays, objects and other things with types) as arguments to builtins and Omni scripts.

Test Plan: Didn't see errors due to `strncmp`, etc. when passing arrays to builtins.

Reviewers: epriestley, #blessed_reviewers

Reviewed By: epriestley, #blessed_reviewers

Subscribers: Korvin, epriestley

Differential Revision: https://secure.phabricator.com/D11138