Prevent error in PhutilArgumentParser when not using standard options
Summary: This prevents an exception from being raised when PhutilArgumentParser is being used without the standard options.  Elsewhere it checks to make sure 'show-standard-options' is an available option before getting it's value, but here it just calls `getArg`, which causes an exception to be raised.

Test Plan: Saw this occur when hitting this code path in Omni (which doesn't have the standard options for builtins).

Reviewers: epriestley, #blessed_reviewers

Reviewed By: epriestley, #blessed_reviewers

Subscribers: Korvin, epriestley

Differential Revision: https://secure.phabricator.com/D11279