Rename XHPAST build script
Summary: Historically XHPAST was built using a shell script, but it is now executed by PHP. We should eventually rename this file for consistency. To retain compatibility, we could possibly symlink `build_xhpast.sh` to `build_xhpast.php` and then check the value of `$argv[0]` and emit a deprecation warning if `build_xhpast.sh` was called.

Test Plan: N/A

Reviewers: #blessed_reviewers, epriestley

Reviewed By: #blessed_reviewers, epriestley

Subscribers: Korvin, epriestley

Differential Revision: https://secure.phabricator.com/D12889